telecom_ai_platform/requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
torch>=2.0.0
joblib>=1.3.0

# API and Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# LangChain and AI
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data Processing
openpyxl>=3.1.0
xlrd>=2.0.1

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
pathlib>=1.0.1

# Development and Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: For enhanced logging
colorlog>=6.7.0
